requirements.txt
requests
pandas
numpy
selenium
pillow
chromedriver-autoinstaller
pytest
allure-pytest
pytest-html
pytest-html-merger